Move modals into their own component
